+++ src/main/java/com/runninghi/bookmark/command/application/service/CommandBookmarkService.java
@@ -1,20 +1,19 @@
 package com.runninghi.bookmark.command.application.service;
 
 import com.runninghi.bookmark.command.application.dto.request.CreateBookmarkRequest;
+import com.runninghi.bookmark.command.application.dto.request.DeleteBookmarkRequest;
 import com.runninghi.bookmark.command.domain.aggregate.entity.Bookmark;
 import com.runninghi.bookmark.command.domain.repository.BookmarkRepository;
 import com.runninghi.bookmark.command.domain.service.BookmarkDomainService;
-import com.runninghi.bookmark.command.infrastructure.service.BookmarkInfraService;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
 
 @Service
 @RequiredArgsConstructor
-public class CreateBookmarkService {
+public class CommandBookmarkService {
 
     private final BookmarkRepository bookmarkRepository;
     private final BookmarkDomainService domainService;
@@ -30,6 +29,14 @@ public class CreateBookmarkService {
                 .userNo(bookmarkDTO.userNo())
                 .addDate(LocalDate.now())
                 .build());
+    }
+
+    @Transactional
+    public void deleteBookmark(DeleteBookmarkRequest bookmarkDTO) {
+
+        domainService.validateFolderExist(bookmarkDTO.bookmarkVO().getFolderNo());
+
+        bookmarkRepository.deleteById(bookmarkDTO.bookmarkVO());
     }
 
 }
