+++ src/main/java/com/runninghi/bookmark/query/application/service/QueryBookmarkService.java
@@ -10,7 +10,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
-public class FindBookmarkService {
+public class QueryBookmarkService {
 
     private final BookmarkRepository bookmarkRepository;
     @Transactional(readOnly=true)
