+++ src/test/java/com/runninghi/bookmark/command/application/service/CommandBookmarkServiceTests.java
@@ -1,9 +1,11 @@
 package com.runninghi.bookmark.command.application.service;
 
 import com.runninghi.bookmark.command.application.dto.request.CreateBookmarkRequest;
+import com.runninghi.bookmark.command.application.dto.request.DeleteBookmarkRequest;
 import com.runninghi.bookmark.command.domain.aggregate.vo.BookmarkVO;
 import com.runninghi.bookmark.command.domain.repository.BookmarkRepository;
-import com.runninghi.bookmarkfolder.command.application.service.CreateNewBookmarkFolderService;
+import com.runninghi.bookmark.query.application.dto.FindBookmarkRequest;
+import com.runninghi.bookmark.query.application.service.QueryBookmarkService;
 import com.runninghi.feedback.command.domain.exception.customException.NotFoundException;
 import jakarta.transaction.Transactional;
 import org.junit.jupiter.api.Assertions;
@@ -17,13 +19,16 @@ import java.util.UUID;
 
 @SpringBootTest
 @Transactional
-public class CreateBookmarkTests {
+public class CommandBookmarkServiceTests {
 
     @Autowired
-    private CreateBookmarkService createBookmarkService;
+    private CommandBookmarkService commandBookmarkService;
 
     @Autowired
     private BookmarkRepository bookmarkRepository;
+
+    @Autowired
+    private QueryBookmarkService queryBookmarkService;
 
 
     @Test
@@ -34,7 +39,7 @@ public class CreateBookmarkTests {
 
         BookmarkVO bookmarkVO = new BookmarkVO(1L, 2L);
         CreateBookmarkRequest bookmarkRequest = new CreateBookmarkRequest(bookmarkVO, UUID.randomUUID());
-        createBookmarkService.createBookmark(bookmarkRequest);
+        commandBookmarkService.createBookmark(bookmarkRequest);
 
         Long afterSize = bookmarkRepository.count();
 
@@ -49,7 +54,7 @@ public class CreateBookmarkTests {
         BookmarkVO bookmarkVO = new BookmarkVO(0L, 1L);
         CreateBookmarkRequest bookmarkRequest = new CreateBookmarkRequest(bookmarkVO, UUID.randomUUID());
 
-        org.assertj.core.api.Assertions.assertThatThrownBy(() -> createBookmarkService.createBookmark(bookmarkRequest))
+        org.assertj.core.api.Assertions.assertThatThrownBy(() -> commandBookmarkService.createBookmark(bookmarkRequest))
                 .isInstanceOf(NotFoundException.class)
                 .hasMessage("해당 폴더가 존재하지 않습니다.");
     }
@@ -64,6 +69,34 @@ public class CreateBookmarkTests {
 //        org.assertj.core.api.Assertions.assertThatThrownBy(() -> createBookmarkService.createBookmark(bookmarkRequest))
 //                .isInstanceOf(NotFoundException.class)
 //                .hasMessage("존재하지 않는 게시물입니다.");
+    }
+
+    @Test
+    @DisplayName("삭제: 즐겨찾기 삭제 기능")
+    void testDeleteBookmark() {
+        BookmarkVO bookmarkVO = new BookmarkVO(1L, 2L);
+        CreateBookmarkRequest createRequest = new CreateBookmarkRequest(bookmarkVO, UUID.randomUUID());
+        commandBookmarkService.createBookmark(createRequest);
+
+        DeleteBookmarkRequest deleteRequest = new DeleteBookmarkRequest(new BookmarkVO(bookmarkVO.getFolderNo(), bookmarkVO.getPostNo()));
+        commandBookmarkService.deleteBookmark(deleteRequest);
+
+        org.assertj.core.api.Assertions.assertThatThrownBy(() -> queryBookmarkService.findBookmark(new FindBookmarkRequest(bookmarkVO)))
+                .isInstanceOf(NotFoundException.class)
+                .hasMessage("존재하지 않는 즐겨찾기 입니다.");
+
+    }
+
+    @Test
+    @DisplayName("삭제: 즐겨찾기 폴더번호 없을 시 예외처리")
+    public void testDeleteFolderNoDoesntExist() {
+
+        DeleteBookmarkRequest deleteRequest = new DeleteBookmarkRequest(new BookmarkVO(0L, 1L));
+
+        Assertions.assertThrows(NotFoundException.class, () -> {
+            commandBookmarkService.deleteBookmark(deleteRequest);
+        });
+
     }
